LPS-82164 Move files to architect package on forms headless
