+++ modules/apps/headless-apio/forms/forms-apio-impl/src/main/java/com/liferay/forms/apio/internal/architect/resource/FormInstanceNestedCollectionResource.java
@@ -35,10 +35,10 @@ import com.liferay.dynamic.data.mapping.service.DDMFormInstanceService;
 import com.liferay.forms.apio.architect.identifier.FormInstanceIdentifier;
 import com.liferay.forms.apio.architect.identifier.StructureIdentifier;
 import com.liferay.forms.apio.internal.architect.form.FormContextForm;
-import com.liferay.forms.apio.internal.form.MediaObjectCreatorForm;
-import com.liferay.forms.apio.internal.representable.EvaluateContextRoute;
-import com.liferay.forms.apio.internal.representable.FormContextIdentifier;
-import com.liferay.forms.apio.internal.representable.FormContextWrapper;
+import com.liferay.forms.apio.internal.architect.form.MediaObjectCreatorForm;
+import com.liferay.forms.apio.internal.architect.route.EvaluateContextRoute;
+import com.liferay.forms.apio.architect.identifier.FormContextIdentifier;
+import com.liferay.forms.apio.internal.model.FormContextWrapper;
 import com.liferay.forms.apio.internal.util.EvaluateContextUtil;
 import com.liferay.forms.apio.internal.util.FormInstanceRepresentorUtil;
 import com.liferay.media.object.apio.architect.identifier.FileEntryIdentifier;
